# Faster B0 shimming by changing some mathematical expressions