## Checklist

- [x] I've given this PR a concise, self-descriptive, and meaningful title
- [x] I've linked relevant issues in the PR body
- [x] I've applied the relevant labels to this PR
- [x] I've added relevant tests for my contribution
- [x] I've updated the documentation and/or added correct docstrings
- [x] I've assigned a reviewer


## Description
This PR is make to be able to achieve faster B0 shimming, indeed, B0 shimming, is an important process in MRI, and need to be fast. However, before this PR, for a dataset like this : 
- coil ( 32 channels) ;
- anatomic image ([174,174,50] );
-  fieldmap([112,112,70]);
-  mask ([112,112,70]). 
The B0 shimming process took around **130 seconds on the Django Computer**, and with multiprocessing. We found that a big part of this time was caused by the optimizations of the different currents which are calculated in the lsq_optimizer class. After some research I found out that a simple modification can speed up a lot the optimization. 
For this, we just have to change some functions with a switch from np.sum(A*B,...), to A@B (or B@A), indeed in python, it's much more faster to do matrix multiplications than np.sum (And it gave exactly the same result ).
Here is the proof : Let's do this line of code : np.sum(coil_mat * coef, axis = 1 ) (This is used in the residuals_mse function)
with dim(coilmat)=(n,m) and dim(coef)=(m)(Here let's call coil_mat A, and coef B)
Also, it's important to note that in python the multiplication of two matrices is different that the mathematical multiplication of matrices.
Here A multiply by B is equal : 
```math
C= \begin{bmatrix}
 A_{1,1}*B_1,  A_{1,2}*B_2,..., A_{1,m}*B_m \\
...,...,...\\
A_{n,1}*B_1, A_{n,2}*B_2,...,A_{n,m}*B_m
\end{bmatrix}
```
And the np.sum(C, axis = 1 ), will do this :
```math
\begin{bmatrix}
 \\
\sum_{i=1}^{m} A_{1,i}*B_i \\
...\\
\sum_{i=1}^{m}A_{n,i}*B_i\\
\end{bmatrix} 
```
(I don't know why the term in the sum are going below, when I'm writing this on github...)
Which is also just the mathematical multiplication of two matrices !
So I've decided to change some parts of the code with that modification.( I didn't change all of the parts where we have np.sum, because it was already fast enough, and it's more readable for a user to keep np.sum(A*B))
This allows having much more faster optimization. Indeed, on Django mac computer, without multiprocessing it is now taking much less time. Indeed it went from **90 seconds to only 20 seconds** (With around 15 of them caused by resampling). So it does helps ! 
I also did a modification in the function eval_static_shim in the sequencer. to reduce the time by around 10 seconds (While keeping the same results). 

With all of these changes, **B0 shimming should now be faster (80 seconds VS 130)**.  Even with the problem of the high time processing of different resampling, and the fact that multiprocessing on mac computer is still really slow to start.(For this reason, I believe that with these change, it's for now better to not use multiprocessing on mac. However, if we succeed in doing faster multiprocessing, the optimization can be way faster !) 

## Linked issues
Adresses https://github.com/shimming-toolbox/shimming-toolbox/pull/423 , https://github.com/shimming-toolbox/shimming-toolbox/issues/412